refactor: 삭제/수정 로직 한 핸들러에 작성
